Split AddressChooserModal.tsx into component and cointainer
